clean up the zuora library
