+++ handlers/catalog-service/src/main/scala/com/gu/catalogService/Handler.scala
@@ -2,15 +2,15 @@ package com.gu.catalogService
 
 import com.amazonaws.services.s3.model.{PutObjectRequest, PutObjectResult}
 import com.gu.effects.RawEffects
-import com.gu.util.reader.Types._
-import com.gu.util.zuora.{ZuoraRestConfig, ZuoraRestRequestMaker}
 import com.gu.util.Logging
 import com.gu.util.config.ConfigReads.ConfigFailure
 import com.gu.util.config.{LoadConfig, Stage, ZuoraEnvironment}
+import com.gu.util.zuora.{ZuoraRestConfig, ZuoraRestRequestMaker}
 import okhttp3.{Request, Response}
 import play.api.libs.json.{Json, Reads}
-import scala.util.Try
 import scalaz.\/
+
+import scala.util.Try
 
 object Handler extends Logging {
