+++ handlers/digital-subscription-expiry/src/test/scala/com/gu/digitalSubscriptionExpiry/DigitalSubscriptionExpiryHandlerEffectsTest.scala
@@ -1,7 +1,9 @@
 package com.gu.digitalSubscriptionExpiry
 
 import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
-import java.time.{LocalDateTime}
+import java.time.LocalDateTime
+
+import com.gu.digitalSubscriptionExpiry.Runner._
 import com.gu.effects.RawEffects
 import com.gu.test.EffectsTest
 import com.gu.util.apigateway.ApiGatewayHandler.LambdaIO
@@ -10,25 +12,16 @@ import play.api.libs.json.Json
 
 class DigitalSubscriptionExpiryHandlerEffectsTest extends FlatSpec with Matchers {
 
-  val now: () => LocalDateTime = () => LocalDateTime.of(2018, 4, 20, 2, 12)
-
   it should "return 404 for invalid subscriber id" taggedAs EffectsTest in {
 
-    val invalidRequest: String =
+    val request: String =
       """
         |{
         |    "body": "{\"subscriberId\" : \"invalidSubId\",\"password\" : \"invalidPassword\"}"
         |}
       """.stripMargin
 
-    val stream = new ByteArrayInputStream(invalidRequest.getBytes(java.nio.charset.StandardCharsets.UTF_8))
-    val os = new ByteArrayOutputStream()
-
-    //execute
-
-    Handler.runWithEffects(RawEffects.createDefault, RawEffects.response, now, LambdaIO(stream, os, null))
-
-    val responseString = new String(os.toByteArray, "UTF-8")
+    val responseString: String = Runner(request)
 
     responseString jsonMatches notFoundResponse
   }
@@ -36,23 +29,17 @@ class DigitalSubscriptionExpiryHandlerEffectsTest extends FlatSpec with Matchers
   it should "successful get expiry date for sub against real backend" taggedAs EffectsTest in {
     val request: String =
       """
-          |{
-          |    "body": "{\"subscriberId\" : \"A-S00044160\",\"password\" : \"W134GH\"}"
-          |}
-        """.stripMargin
+        |{
+        |    "body": "{\"subscriberId\" : \"A-S00044160\",\"password\" : \"W134GH\"}"
+        |}
+      """.stripMargin
 
-    val stream = new ByteArrayInputStream(request.getBytes(java.nio.charset.StandardCharsets.UTF_8))
-    val os = new ByteArrayOutputStream()
-
-    //execute
-    Handler.runWithEffects(RawEffects.createDefault, RawEffects.response, now, LambdaIO(stream, os, null))
-
-    val responseString = new String(os.toByteArray, "UTF-8")
+    val responseString: String = Runner(request)
 
     val expectedResponse =
       """
-             |{"statusCode":"200","headers":{"Content-Type":"application/json"},"body":"{\n  \"expiry\" : {\n    \"expiryDate\" : \"2018-11-30\",\n    \"expiryType\" : \"sub\",\n    \"content\" : \"SevenDay\"\n  }\n}"}
-             |""".stripMargin
+        |{"statusCode":"200","headers":{"Content-Type":"application/json"},"body":"{\n  \"expiry\" : {\n    \"expiryDate\" : \"2018-11-30\",\n    \"expiryType\" : \"sub\",\n    \"content\" : \"SevenDay\"\n  }\n}"}
+        |""".stripMargin
     responseString jsonMatches expectedResponse
   }
 
@@ -64,13 +51,7 @@ class DigitalSubscriptionExpiryHandlerEffectsTest extends FlatSpec with Matchers
         |}
       """.stripMargin
 
-    val stream = new ByteArrayInputStream(request.getBytes(java.nio.charset.StandardCharsets.UTF_8))
-    val os = new ByteArrayOutputStream()
-
-    //execute
-    Handler.runWithEffects(RawEffects.createDefault, RawEffects.response, now, LambdaIO(stream, os, null))
-
-    val responseString = new String(os.toByteArray, "UTF-8")
+    val responseString: String = Runner(request)
 
     val expectedResponse =
       """
@@ -78,7 +59,24 @@ class DigitalSubscriptionExpiryHandlerEffectsTest extends FlatSpec with Matchers
         |""".stripMargin
     responseString jsonMatches expectedResponse
   }
-  implicit class JsonMatcher(private val actual: String) {
+
+}
+object Runner {
+
+  val now: () => LocalDateTime = () => LocalDateTime.of(2018, 4, 20, 2, 12)
+
+  def apply(request: String) = {
+    val stream = new ByteArrayInputStream(request.getBytes(java.nio.charset.StandardCharsets.UTF_8))
+    val os = new ByteArrayOutputStream()
+
+    //execute
+    Handler.runWithEffects(RawEffects.stage, RawEffects.s3Load, RawEffects.response, now, LambdaIO(stream, os, null))
+
+    val responseString = new String(os.toByteArray, "UTF-8")
+    responseString
+  }
+
+  implicit class JsonMatcher(private val actual: String) extends Matchers {
     def jsonMatches(expected: String): Assertion = {
       val expectedJson = Json.parse(expected)
       val actualJson = Json.parse(actual)
