+++ handlers/digital-subscription-expiry/src/test/scala/com/gu/digitalSubscriptionExpiry/emergencyToken/GetTokenExpiryTest.scala
@@ -1,9 +1,10 @@
 package com.gu.digitalSubscriptionExpiry.emergencyToken
 
 import java.time.LocalDate
+
 import com.gu.cas.{PrefixedTokens, SevenDay}
 import com.gu.digitalSubscriptionExpiry.responses.{Expiry, ExpiryType, SuccessResponse}
-import com.gu.util.apigateway.ResponseModels.{ApiResponse, Headers}
+import com.gu.util.apigateway.ResponseModels.ApiResponse
 import org.scalatest.{FlatSpec, Matchers}
 import play.api.libs.json.Json
 import scalaz.{-\/, \/-}
