+++ handlers/identity-backfill/src/main/scala/com/gu/identityBackfill/zuora/AddIdentityIdToAccount.scala
@@ -3,8 +3,7 @@ package com.gu.identityBackfill.zuora
 import com.gu.identityBackfill.Types
 import com.gu.identityBackfill.Types.{AccountId, IdentityId}
 import com.gu.util.zuora.RestRequestMaker.{ClientFailableOp, Requests}
-import com.gu.util.zuora.ZuoraReaders.unitReads
-import play.api.libs.json.Json
+import play.api.libs.json.{JsSuccess, Json, Reads}
 
 object AddIdentityIdToAccount {
 
@@ -14,6 +13,9 @@ object AddIdentityIdToAccount {
   def reqFromIdentityId(id: Types.IdentityId): WireRequest = {
     WireRequest(id.value)
   }
+
+  implicit val unitReads: Reads[Unit] =
+    Reads(_ => JsSuccess(()))
 
   def apply(requests: Requests)(accountId: AccountId, identityId: IdentityId): ClientFailableOp[Unit] =
     requests.put[WireRequest, Unit](reqFromIdentityId(identityId), s"accounts/${accountId.value}")
