+++ handlers/identity-backfill/src/test/scala/com/gu/identityBackfill/EndToEndHandlerTest.scala
@@ -74,7 +74,7 @@ object Runner {
     val config = new TestingRawEffects(false, 200, responses, postResponses)
 
     //execute
-    Handler.runWithEffects(config.rawEffects, config.response, LambdaIO(stream, os, null))
+    Handler.runWithEffects(config.stage, config.s3Load, config.response, LambdaIO(stream, os, null))
 
     val responseString = new String(os.toByteArray, "UTF-8")
