+++ handlers/identity-backfill/src/test/scala/com/gu/identityBackfill/salesforce/getContact/GetSFContactSyncCheckFieldsEffectsTest.scala
@@ -15,7 +15,7 @@ class GetSFContactSyncCheckFieldsEffectsTest extends FlatSpec with Matchers {
     val testContact = SFContactId("003g000000LEwO6AAL")
 
     val actual = for {
-      auth <- DevSFEffects(RawEffects.createDefault, RawEffects.response)
+      auth <- DevSFEffects(RawEffects.s3Load, RawEffects.response)
       authed = GetSFContactSyncCheckFields(auth) _
       result <- authed(testContact)
     } yield result
