+++ src/main/scala/com/gu/stripeCustomerSourceUpdated/zuora/CreatePaymentMethod.scala
@@ -4,7 +4,7 @@ import com.gu.stripeCustomerSourceUpdated._
 import com.gu.util.ZuoraToApiGateway
 import com.gu.util.reader.Types._
 import com.gu.util.zuora.RestRequestMaker.Requests
-import com.gu.util.zuora.ZuoraAccount.{AccountId, NumConsecutiveFailures, PaymentMethodId}
+import com.gu.util.zuora.ZuoraGetAccountSummary.ZuoraAccount.{AccountId, NumConsecutiveFailures, PaymentMethodId}
 import play.api.libs.json.{JsPath, Json, Reads, Writes}
 
 object CreatePaymentMethod {
