+++ src/main/scala/com/gu/stripeCustomerSourceUpdated/zuora/SetDefaultPaymentMethod.scala
@@ -3,9 +3,8 @@ package com.gu.stripeCustomerSourceUpdated.zuora
 import com.gu.util.ZuoraToApiGateway
 import com.gu.util.reader.Types._
 import com.gu.util.zuora.RestRequestMaker.Requests
-import com.gu.util.zuora.ZuoraAccount.{AccountId, PaymentMethodId}
-import com.gu.util.zuora.ZuoraReaders.unitReads
-import play.api.libs.json.{Json, Writes}
+import com.gu.util.zuora.ZuoraGetAccountSummary.ZuoraAccount.{AccountId, PaymentMethodId}
+import play.api.libs.json.{JsSuccess, Json, Reads, Writes}
 
 object SetDefaultPaymentMethod {
 
@@ -17,6 +16,9 @@ object SetDefaultPaymentMethod {
       "DefaultPaymentMethodId" -> subscriptionUpdate.paymentMethodId
     )
   }
+
+  implicit val unitReads: Reads[Unit] =
+    Reads(_ => JsSuccess(()))
 
   def setDefaultPaymentMethod(requests: Requests)(accountId: AccountId, paymentMethodId: PaymentMethodId): FailableOp[Unit] =
     requests.put(SetDefaultPaymentMethod(paymentMethodId), s"object/account/${accountId.value}").leftMap(ZuoraToApiGateway.fromClientFail)
