+++ src/main/scala/com/gu/util/zuora/ZuoraDisableAutoPay.scala
@@ -1,0 +1,22 @@
+package com.gu.util.zuora
+
+import com.gu.util.zuora.RestRequestMaker.{ClientFailableOp, Requests}
+import play.api.libs.json.{JsSuccess, Json, Reads, Writes}
+
+object ZuoraDisableAutoPay {
+
+  case class AccountUpdate(autoPay: Boolean)
+
+  implicit val accountUpdateWrites = new Writes[AccountUpdate] {
+    def writes(accountUpdate: AccountUpdate) = Json.obj(
+      "autoPay" -> accountUpdate.autoPay
+    )
+  }
+
+  implicit val unitReads: Reads[Unit] =
+    Reads(_ => JsSuccess(()))
+
+  def apply(requests: Requests)(accountId: String): ClientFailableOp[Unit] =
+    requests.put(AccountUpdate(autoPay = false), s"accounts/$accountId"): ClientFailableOp[Unit]
+
+}
