+++ src/main/scala/com/gu/util/zuora/ZuoraGetInvoiceTransactions.scala
@@ -1,0 +1,43 @@
+package com.gu.util.zuora
+
+import java.time.LocalDate
+
+import com.gu.util.zuora.RestRequestMaker.{ClientFailableOp, Requests}
+import play.api.libs.json.{JsPath, Reads}
+import play.api.libs.functional.syntax._
+
+object ZuoraGetInvoiceTransactions {
+  case class InvoiceItem(id: String, subscriptionName: String, serviceStartDate: LocalDate, serviceEndDate: LocalDate, chargeAmount: Double, chargeName: String, productName: String)
+
+  case class ItemisedInvoice(id: String, invoiceDate: LocalDate, amount: Double, balance: Double, status: String, invoiceItems: List[InvoiceItem])
+
+  case class InvoiceTransactionSummary(invoices: List[ItemisedInvoice])
+
+  implicit val invoiceItemReads: Reads[InvoiceItem] = (
+    (JsPath \ "id").read[String] and
+    (JsPath \ "subscriptionName").read[String] and
+    (JsPath \ "serviceStartDate").read[LocalDate] and
+    (JsPath \ "serviceEndDate").read[LocalDate] and
+    (JsPath \ "chargeAmount").read[Double] and
+    (JsPath \ "chargeName").read[String] and
+    (JsPath \ "productName").read[String]
+  )(InvoiceItem.apply _)
+
+  implicit val itemisedInvoiceReads: Reads[ItemisedInvoice] = (
+    (JsPath \ "id").read[String] and
+    (JsPath \ "invoiceDate").read[LocalDate] and
+    (JsPath \ "amount").read[Double] and
+    (JsPath \ "balance").read[Double] and
+    (JsPath \ "status").read[String] and
+    (JsPath \ "invoiceItems").read[List[InvoiceItem]]
+  )(ItemisedInvoice.apply _)
+
+  implicit val invoiceTransactionSummaryReads: Reads[InvoiceTransactionSummary] =
+    (JsPath \ "invoices").read[List[ItemisedInvoice]].map {
+      invoices => InvoiceTransactionSummary(invoices)
+    }
+
+  def apply(requests: Requests)(accountId: String): ClientFailableOp[InvoiceTransactionSummary] =
+    requests.get[InvoiceTransactionSummary](s"transactions/invoices/accounts/$accountId")
+
+}
