+++ src/main/scala/com/gu/util/zuora/ZuoraUpdateCancellationReason.scala
@@ -1,0 +1,23 @@
+package com.gu.util.zuora
+
+import com.gu.util.zuora.RestRequestMaker.{ClientFailableOp, Requests}
+import com.gu.util.zuora.ZuoraGetAccountSummary.SubscriptionId
+import play.api.libs.json.{JsSuccess, Json, Reads, Writes}
+
+object ZuoraUpdateCancellationReason {
+
+  case class SubscriptionUpdate(cancellationReason: String)
+
+  implicit val subscriptionUpdateWrites = new Writes[SubscriptionUpdate] {
+    def writes(subscriptionUpdate: SubscriptionUpdate) = Json.obj(
+      "CancellationReason__c" -> subscriptionUpdate.cancellationReason
+    )
+  }
+
+  implicit val unitReads: Reads[Unit] =
+    Reads(_ => JsSuccess(()))
+
+  def apply(requests: Requests)(subscription: SubscriptionId): ClientFailableOp[Unit] =
+    requests.put(SubscriptionUpdate("System AutoCancel"), s"subscriptions/${subscription.id}"): ClientFailableOp[Unit]
+
+}
