+++ src/test/scala/com/gu/paymentFailure/PaymentFailureHandlerTest.scala
@@ -5,16 +5,14 @@ import java.io.ByteArrayOutputStream
 import com.gu.TestData
 import com.gu.TestData._
 import com.gu.stripeCustomerSourceUpdated.SourceUpdatedSteps.StepsConfig
-import com.gu.util.config.ETConfig.ETSendId
 import com.gu.util.apigateway.ApiGatewayHandler.{LambdaIO, Operation}
-import com.gu.util.apigateway.ApiGatewayResponse.ResponseBody
 import com.gu.util.apigateway.{ApiGatewayHandler, ApiGatewayResponse}
+import com.gu.util.config.ETConfig.ETSendId
 import com.gu.util.config.Stage
 import com.gu.util.exacttarget._
 import com.gu.util.reader.Types._
 import com.gu.util.zuora.ZuoraGetInvoiceTransactions.InvoiceTransactionSummary
 import org.scalatest.{FlatSpec, Matchers}
-
 import scalaz.{-\/, \/-}
 
 class PaymentFailureHandlerTest extends FlatSpec with Matchers {
